Make use of new django-celery-model library

+++ webgcal/apps/webgcal/views.py
@@ -92,7 +92,10 @@ def create_calendar(request):
 
 @login_required
 def edit_calendar(request, calendar_id):
-    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id, task_id=None)
+    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
+    if calendar.running:
+        return HttpResponseForbidden('Cannot manipulate calendar with active background task.')
+
     edit_form = CalendarForm(instance=calendar, data=request.POST if request.method == 'POST' else None)
 
     if edit_form.is_valid():
@@ -113,7 +116,10 @@ def edit_calendar(request, calendar_id):
 
 @login_required
 def switch_calendar(request, calendar_id):
-    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id, task_id=None)
+    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
+    if calendar.running:
+        return HttpResponseForbidden('Cannot manipulate calendar with active background task.')
+
     calendar.enabled = not(calendar.enabled)
     calendar.save()
 
@@ -123,7 +129,10 @@ def switch_calendar(request, calendar_id):
 
 @login_required
 def delete_calendar(request, calendar_id):
-    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id, task_id=None)
+    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
+    if calendar.running:
+        return HttpResponseForbidden('Cannot manipulate calendar with active background task.')
+
     calendar.delete()
 
     messages.success(request, 'Deleted calendar "%s" from your Dashboard!' % calendar)
@@ -133,7 +142,10 @@ def delete_calendar(request, calendar_id):
 @login_required
 def delete_calendar_ask(request, calendar_id):
     calendars = Calendar.objects.filter(user=request.user).order_by('name')
-    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id, task_id=None)
+    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
+    if calendar.running:
+        return HttpResponseForbidden('Cannot manipulate calendar with active background task.')
+
     create_form = CalendarForm()
 
     button = '<a class="ym-button ym-delete float-right" href="%s" title="Yes">Yes</a>' % reverse('webgcal:delete_calendar', kwargs={'calendar_id': calendar_id})
@@ -148,11 +160,13 @@ def delete_calendar_ask(request, calendar_id):
 
 @permission_required('webgcal.edit_calendar')
 def sync_calendar_now(request, calendar_id):
-    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id, task_id=None)
-    calendar.task_id = 'sync-calendar-%d-%d' % (request.user.id, calendar.id)
-    calendar.save()
+    calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
+    if calendar.running:
+        return HttpResponseForbidden('Cannot manipulate calendar with active background task.')
 
-    task_sync_calendar.apply_async(args=[request.user.id, calendar.id], task_id=calendar.task_id)
+    args = (request.user.id, calendar.id)
+    task_id = 'sync-calendar-%d-%d' % args
+    calendar.apply_async(task_sync_calendar, args=args, task_id=task_id, countdown=10)
 
     messages.info(request, 'Queued syncing of calendar "%s" ...' % calendar)
 
@@ -183,7 +197,10 @@ def create_website(request, calendar_id):
 @login_required
 def edit_website(request, calendar_id, website_id):
     calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
-    website = get_object_or_404(Website, calendar=calendar, id=website_id, task_id=None)
+    website = get_object_or_404(Website, calendar=calendar, id=website_id)
+    if website.running:
+        return HttpResponseForbidden('Cannot manipulate website with active background task.')
+
     edit_form = WebsiteForm(instance=website, data=request.POST if request.method == 'POST' else None)
 
     if edit_form.is_valid():
@@ -206,7 +223,10 @@ def edit_website(request, calendar_id, website_id):
 @login_required
 def switch_website(request, calendar_id, website_id):
     calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
-    website = get_object_or_404(Website, calendar=calendar, id=website_id, task_id=None)
+    website = get_object_or_404(Website, calendar=calendar, id=website_id)
+    if website.running:
+        return HttpResponseForbidden('Cannot manipulate website with active background task.')
+
     website.enabled = not(website.enabled)
     website.save()
 
@@ -217,7 +237,10 @@ def switch_website(request, calendar_id, website_id):
 @login_required
 def delete_website(request, calendar_id, website_id):
     calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
-    website = get_object_or_404(Website, calendar=calendar, id=website_id, task_id=None)
+    website = get_object_or_404(Website, calendar=calendar, id=website_id)
+    if website.running:
+        return HttpResponseForbidden('Cannot manipulate website with active background task.')
+
     website.delete()
 
     messages.success(request, 'Deleted website "%s" from your Dashboard!' % website)
@@ -228,7 +251,10 @@ def delete_website(request, calendar_id, website_id):
 def delete_website_ask(request, calendar_id, website_id):
     calendars = Calendar.objects.filter(user=request.user).order_by('name')
     calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
-    website = get_object_or_404(Website, calendar=calendar, id=website_id, task_id=None)
+    website = get_object_or_404(Website, calendar=calendar, id=website_id)
+    if website.running:
+        return HttpResponseForbidden('Cannot manipulate website with active background task.')
+
     create_form = WebsiteForm()
 
     button = '<a class="ym-button ym-delete float-right" href="%s" title="Yes">Yes</a>' % reverse('webgcal:delete_website', kwargs={'calendar_id': calendar_id, 'website_id': website_id})
@@ -245,11 +271,13 @@ def delete_website_ask(request, calendar_id, website_id):
 @permission_required('webgcal.edit_website')
 def parse_website_now(request, calendar_id, website_id):
     calendar = get_object_or_404(Calendar, user=request.user, id=calendar_id)
-    website = get_object_or_404(Website, calendar=calendar, id=website_id, task_id=None)
-    website.task_id = 'parse-website-%d-%d' % (request.user.id, website.id)
-    website.save()
+    website = get_object_or_404(Website, calendar=calendar, id=website_id)
+    if website.running:
+        return HttpResponseForbidden('Cannot manipulate website with active background task.')
 
-    task_parse_website.apply_async(args=[request.user.id, website.id], task_id=website.task_id)
+    args = (request.user.id, website.id)
+    task_id = 'parse-website-%d-%d' % args
+    website.apply_async(task_parse_website, args=args, task_id=task_id, countdown=10)
 
     messages.info(request, 'Queued parsing of website "%s" ...' % website)
